EA-9892: Update mod_security2 from v2.9.3 to v2.9.4
Patch fixup.
Patch 0002 is no longer applicable as the files it patches
are no longer included in the tar.gz file.  They are supposed
to be generated by autoconf.
Patch 0003 is now in part of 2.9.4:
https://github.com/SpiderLabs/ModSecurity/commit/e419b50fe7051245adc7d5d230eb2dd951343df8
patch 0004 is also in 2.9.4:
https://github.com/SpiderLabs/ModSecurity/commit/52532a1bce0b705c0aa4365fecf727b836d37f00

Fixed the spec file to run autogen.sh to generate the necessary
build files and added update the BuildRequires to include some
other packages now required for the build.