EA-10239: Increase compiler optimization level to help mitigate performance loss from large rulesets
By appending -O3 to "optflags", we ensure that CFLAGS is set with
its usual values, but with -O3 appended.  Appending -O3 will override
the -O2 option that is set by default.
From the gcc man page:
"If you use multiple -O options, with or without level numbers,
the last such option is the one that is effective."